Made pandas, matplotlib, ipython and scikit-image extras.
This change is backward incompatible: installation now requires
[datasets,plots] to work as before.

This change is necessary to avoid installing many packages that
are not needed to compute shap values, just to plot them or use
pre-built datasets. These dependencies may not be available
on architectures without access to the internet (e.g. proxied pypi).

This change makes shap only depend on:
['numpy', 'scipy', 'scikit-learn', 'tqdm'].